M    scripts/so_epilogue.mk M    scripts/program.mk M    configVars.mk.in
Allowed for 'quiet' build.  If you set the environment variable
SCI_MAKE_BE_QUIET to 'true', then instead of the entire compile line
being printed, only the name of the file will be shown.  By default,
this behavior is off and you won't see it, but I suggest you give it a
try.  If you run into compilation errors where you need to see the
compile line, you can just unset that environment variable.

M    Packages/Uintah/CCA/Components/ICE/AMRICE.h
M    Packages/Uintah/CCA/Components/ICE/impAMRICE.cc
M    Packages/Uintah/CCA/Components/ICE/AMRICE.cc

Moved template files from AMRICE.cc to AMRICE.h... this fixed a
problem (on the SGI) where impAMRICE.cc needed certain template
functions and thus the compiler (CC) pulled in ICE.cc (don't know why
it was ICE.cc, but it was).  This caused compilation errors.  By
moving the template files to that .h file, this problem goes away, yet
the organization of the functions still remains good.

M    Packages/Uintah/StandAlone/sub.mk

Uda2nrrd depends on Core/Basis...

+++ src/scripts/program.mk
@@ -64,8 +64,15 @@ $(PROGRAM)_LIBS := $(LIBS)
 # lib/libCore_Thread.so to -lCore_Thread.  This is so that
 # we can use the -l syntax to link, but still express the dependicies.
 $(PROGRAM): $(OBJS) $(patsubst %,$(LIBDIR)/lib%.$(SO_OR_A_FILE),$(PSELIBS))
+ifeq ($(SCI_MAKE_BE_QUIET),true)
+	@rm -f $@
+	@echo "Building:  $@"
+	@$(CXX) $(PROGRAM_LDFLAGS) $(SCI_THIRDPARTY_LIBRARY) $(LDFLAGS) $(LDRUN_PREFIX)$(LIBDIR_ABS) -o $@ $(filter %.o,$^) $(patsubst $(LIBDIR)/lib%.$(SO_OR_A_FILE),-l%,$(filter %.$(SO_OR_A_FILE),$^)) $(REPOSITORIES_$@) $($@_LIBS) $(TAU_LIBRARY)
+else
 	rm -f $@
 	$(CXX) $(PROGRAM_LDFLAGS) $(SCI_THIRDPARTY_LIBRARY) $(LDFLAGS) $(LDRUN_PREFIX)$(LIBDIR_ABS) -o $@ $(filter %.o,$^) $(patsubst $(LIBDIR)/lib%.$(SO_OR_A_FILE),-l%,$(filter %.$(SO_OR_A_FILE),$^)) $(REPOSITORIES_$@) $($@_LIBS) $(TAU_LIBRARY)
+endif
+
 
 #  These will get removed on make clean
 CLEANOBJS := $(CLEANOBJS) $(OBJS)
