+++ src/scripts/so_epilogue.mk
@@ -89,19 +89,33 @@ TMP_PACK_PSELIBS = $(patsubst PACKAGE%,libPackages_%,$(TMP))
 PACK_PSELIBS = $(patsubst SCIRUN%,,$(TMP_PACK_PSELIBS))
 
 $(LIBNAME): $(OBJS) $(patsubst %,$(SCIRUN_LIBDIR)/%,$(CORE_PSELIBS)) $(patsubst %,$(LIBDIR)/%,$(PACK_PSELIBS))
+  ifeq ($(SCI_MAKE_BE_QUIET),true)
+	@rm -f $@
+  else
 	rm -f $@
+  endif
   ifeq ($(IS_AIX),yes)
 	ar -v -q $@ $(filter %.o,$^)
   else
-  ifeq ($(IS_OSX),yes)
+    ifeq ($(IS_OSX),yes)
+      ifeq ($(SCI_MAKE_BE_QUIET),true)
+	@echo "Linking:  $@"
+	@$(CXX) $(SCI_THIRDPARTY_LIBRARY) -single_module $(LDFLAGS) $(SOFLAGS) -install_name $(SCIRUN_LIBDIR_ABS)/$(patsubst lib/%,%,$@) -o $@ $(SONAMEFLAG) $(filter %.o,$^) $(patsubst $(SCIRUN_LIBDIR)/lib%.so,-l%,$(filter %.$(SO_OR_A_FILE),$^)) $(REPOSITORIES_$@) $($(notdir $@)_LIBS) $(TAU_MPI_LIBS) $(TAU_SHLIBS)
+      else
 	$(CXX) $(SCI_THIRDPARTY_LIBRARY) -single_module $(LDFLAGS) $(SOFLAGS) -install_name $(SCIRUN_LIBDIR_ABS)/$(patsubst lib/%,%,$@) -o $@ $(SONAMEFLAG) $(filter %.o,$^) $(patsubst $(SCIRUN_LIBDIR)/lib%.so,-l%,$(filter %.$(SO_OR_A_FILE),$^)) $(REPOSITORIES_$@) $($(notdir $@)_LIBS) $(TAU_MPI_LIBS) $(TAU_SHLIBS)
-  else
-  ifeq ($(IS_WIN),yes)
+      endif
+    else
+      ifeq ($(IS_WIN),yes)
 	$(CXX) -o $@ $(SONAMEFLAG) $(filter %.o,$^) $(patsubst $(SCIRUN_LIBDIR)/lib%.$(SO_OR_A_FILE),-l%,$(filter %.$(SO_OR_A_FILE),$^)) $(REPOSITORIES_$@) $($(notdir $@)_LIBS) $(TAU_MPI_LIBS) $(TAU_SHLIBS) $(LDFLAGS) $(SOFLAGS) $(SCI_THIRDPARTY_LIBRARY) -Wl,--output-def=$(subst dll,def,$@) -Wl,--export-all-symbols -Wl,--out-implib=$(subst dll,a,$@) -Wl,--enable-auto-image-base
-  else
+      else
+        ifeq ($(SCI_MAKE_BE_QUIET),true)
+	@echo "Linking:   $@"
+	@$(CXX) $(SCI_THIRDPARTY_LIBRARY) $(LDFLAGS) $(SOFLAGS) $(LDRUN_PREFIX)$(LIBDIR_ABS) $(LDRUN_PREFIX)$(SCIRUN_LIBDIR_ABS) -o $@ $(SONAMEFLAG) $(filter %.o,$^) $(patsubst $(SCIRUN_LIBDIR)/lib%.so,-l%,$(filter %.$(SO_OR_A_FILE),$^)) $(REPOSITORIES_$@) $($(notdir $@)_LIBS) $(TAU_MPI_LIBS) $(TAU_SHLIBS)
+        else
 	$(CXX) $(SCI_THIRDPARTY_LIBRARY) $(LDFLAGS) $(SOFLAGS) $(LDRUN_PREFIX)$(LIBDIR_ABS) $(LDRUN_PREFIX)$(SCIRUN_LIBDIR_ABS) -o $@ $(SONAMEFLAG) $(filter %.o,$^) $(patsubst $(SCIRUN_LIBDIR)/lib%.so,-l%,$(filter %.$(SO_OR_A_FILE),$^)) $(REPOSITORIES_$@) $($(notdir $@)_LIBS) $(TAU_MPI_LIBS) $(TAU_SHLIBS)
-  endif
-  endif
+        endif
+      endif
+    endif
   endif
 
 #$(LIBNAME).pure: $(LIBNAME)
